Disable all drivers while at least one is disabled
A driver is torque-enabled only when its status word is 55. When all
drivers are torque-enabled, the system state is switched to
operation-enabled. If any driver is no longer torque-enabled, the
system state is switched to idle, yet the remaining drivers may
still be torque-enabled due to the previous implementation. This is
not aligned with the overall design of the state controller.

Set control word to 0 to disable all drivers if any of the is no
longer torque-enabled.